border scale

+++ unity_project/Assets/Scripts/Health.cs
@@ -3,7 +3,7 @@
 public interface Health
 {
 
-		void takeDamage (int damage, string attacker);
+		void takeDamage (float damage, string attacker);
 		void Destroy ();
 		string lastDamageBy ();
 		float HP {
